refactor(pipeline): split stages.go into focused domain files
stages.go (401 lines, all stage logic) → 4 focused files:
- stages.go: pipelineCtx struct, cache helpers, stageValidate
- stages_engine.go: engine build dispatch + container/native builders
- stages_game.go: server + client build dispatch + all game builders
- stages_deploy.go: container build/push, deploy, session, state helpers

stageEngineBuild (CC 6) and stageGameBuild (CC 7) reduced to CC 3 each
by extracting dispatchEngineBuild/dispatchGameBuild — matching the
existing stageClientBuild pattern. No behavior or API changes.
